eric-eo-lm-consumer-lm-integration/src/main/java/com/ericsson/licenseconsumer/lm/integration/util/JsonUtil.java
/*******************************************************************************
 * COPYRIGHT Ericsson 2024
 *
 *
 *
 * The copyright to the computer program(s) herein is the property of
 *
 * Ericsson Inc. The programs may be used and/or copied only with written
 *
 * permission from Ericsson Inc. or in accordance with the terms and
 *
 * conditions stipulated in the agreement/contract under which the
 *
 * program(s) have been supplied.
 ******************************************************************************/
package com.ericsson.licenseconsumer.lm.integration.util;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;

@Component
public class JsonUtil {
    private final ObjectMapper objectMapper;

    @Autowired
    public JsonUtil(ObjectMapper objectMapper) {
        this.objectMapper = objectMapper;
    }

    public <T> T deserializeJson(String json, Class<T> classReference) throws JsonProcessingException {
        return objectMapper.readValue(json, classReference);
    }

    public <T> String serializeJson(final T productTypeRequest) throws JsonProcessingException {
        return objectMapper.writeValueAsString(productTypeRequest);
    }
}
